Used netiface to get self mac address
I added a function that gets all network interfaces with the netifaces package. It loops through all interfaces and checks for a MAC address. It de-concantenates it if needed and alse verifies that it is not 0:0:0:0 If no mac address is found, an OSerror is raised. Else, it returns a list of mac addresses.